backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
solana>=0.34.3
requests>=2.31.0
